Align the voice records and guide with the shipped interview seams
ADR-0009 and its summaries still described a generic composer seam only
and a microphone that closes whenever Brunch handles a turn, which the
interview stage and one-answer buffering contradict. Record the shipped
seams, note that the composer control remains supported, describe the new
transcript and recovery wording in the user guide, and move the shared
sweep output schema out of the local-storage demo folder.
